requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
DAWG-Python==0.7.2
dbfread==2.0.7
dnspython==2.0.0
docopt==0.6.2
email-validator==1.1.2
et-xmlfile==1.0.1
google-search-results-serpwow==1.1.11
googlesearch-python==2020.0.2
idna==2.10
jdcal==1.4.1
joblib==0.17.0
lxml==4.6.2
nltk==3.5
numpy==1.19.4
openpyxl==3.0.6
pandas==1.1.4
phonenumbers==8.12.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
tqdm==4.56.0
urllib3==1.25.9
xlrd==2.0.1
